Wins counter for "old ai vs new ai" mode
Also contains a bugfix for number of rounds per organism for the old vs
new training mode ("neuroevolution ai 2"), and a restructuring of
mutation settings out of the mutation function and into the settings
file.

AND ignores the /old/ folder of stuff I have. Ideally the .gitignore
file itself would be ignored too, because no one else needs it... But
that doesn't seem like a thing.

Yes, this is a lot of changes in one commit. Yes, I should commit more
incremental changes.