Added population win/losses to export data
With gamemodes that don't store win/loss data, the headings exist but
there are no values put in. Not a major issue, but it would be better if
either the headings were removed, or replaced with a "no win/loss data"
indicator.

There's also a bug with exporting the mutations settings (it just
exports "[Object object]"). Maybe it could be JSON.stringified and/or
turned into a string somehow.